// src/game/TimePilot.types.ts
export interface Coordinates {
  posX: number;
  posY: number;
}
// ...
export interface AssetProgress {
  loaded: number;
  remaining: number;
}
// ...
export interface Controller {
  disconnect?: () => void;
}